+++ test/runtests.jl
@@ -49,6 +49,8 @@ if TEST_MODE in ("unit", "all")
 		include("./component/test_vasp_to_extxyz.jl")
 		include("./component/test_poscar_to_toml.jl")
 		include("./component/test_oszicar_to_embset.jl")
+		include("./component/test_IncarIO.jl")
+		include("./component/test_MfaSampling.jl")
 		include("./component/test_sunny_export.jl")
 		include("./component/test_thread_safety.jl")
 	end